# public_docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This API provides methods to create, read and delete shortened URLs.\nThe shortened URLs are generated by creating a hash of the timestamp when they are requested the first time and stored on Mongo.\nFollowing requests for the same URL will return the short url that was previously stored.\nThe API also implements a counter of the shortened url redirections. Each time a short URL performs a redirect, the counter is incremented.",
    "title": "ShortToMe",
    "contact": {
      "name": "Giuseppe Siragusa",
      "url": "https://www.linkedin.com/in/gsiragusa/",
      "email": "giuseppe.siragusa85@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8081",
  "basePath": "/",
  "paths": {
    "/api": {
      "get": {
        "description": "Receives a short url and returns the extended one",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Api"
        ],
        "summary": "Read short url",
        "operationId": "readShortUrl",
        "parameters": [
          {
            "type": "string",
            "description": "short url to read",
            "name": "url",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Extended url",
            "schema": {
              "type": "object",
              "properties": {
                "operation": {
                  "type": "string",
                  "example": "read"
                },
                "status": {
                  "type": "string",
                  "example": "ok"
                },
                "url": {
                  "type": "string",
                  "example": "https://www.google.com"
                }
              }
            }
          },
          "400": {
            "description": "Not Found"
          },
          "404": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Consumes a url and shortens it",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Api"
        ],
        "summary": "Shorten url",
        "operationId": "createShortUrl",
        "parameters": [
          {
            "type": "string",
            "description": "url to shorten",
            "name": "url",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Short url",
            "schema": {
              "type": "object",
              "properties": {
                "operation": {
                  "type": "string",
                  "example": "create"
                },
                "status": {
                  "type": "string",
                  "example": "ok"
                },
                "url": {
                  "type": "string",
                  "example": "http://www.example.com/RMAp1Vz"
                }
              }
            }
          },
          "400": {
            "description": "Not Found"
          },
          "404": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "Deletes the short url",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Api"
        ],
        "summary": "Delete short url",
        "operationId": "deleteShortUrl",
        "parameters": [
          {
            "type": "string",
            "description": "short url to delete",
            "name": "url",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Operation result",
            "schema": {
              "type": "object",
              "properties": {
                "operation": {
                  "type": "string",
                  "example": "delete"
                },
                "status": {
                  "type": "string",
                  "example": "ok"
                },
                "url": {
                  "type": "string",
                  "example": "http://www.example.com/RMAp1Vz"
                }
              }
            }
          },
          "400": {
            "description": "Not Found"
          },
          "404": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/count": {
      "get": {
        "description": "Returns the number of redirections for a given short url",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Api"
        ],
        "summary": "Count number of redirects",
        "operationId": "countRedirects",
        "parameters": [
          {
            "type": "string",
            "description": "short url to count redirections",
            "name": "url",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Total count",
            "schema": {
              "type": "object",
              "properties": {
                "operation": {
                  "type": "string",
                  "example": "count"
                },
                "status": {
                  "type": "string",
                  "example": "ok"
                },
                "url": {
                  "type": "integer",
                  "example": 2
                }
              }
            }
          },
          "400": {
            "description": "Not Found"
          },
          "404": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{shortId}": {
      "get": {
        "description": "Redirect to extended url",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Redirect"
        ],
        "summary": "Redirect to extended url",
        "operationId": "countRedirects",
        "parameters": [
          {
            "type": "string",
            "description": "the id of the short url for redirection",
            "name": "shortId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "301": {
            "description": "Redirects to extended url"
          },
          "400": {
            "description": "Not Found"
          },
          "404": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}